% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_ORA.R
\name{genes_ORA}
\alias{genes_ORA}
\title{Perform over-representation analysis (ORA) using a hypergeometric test}
\usage{
genes_ORA(genes_drawn, universe, database, min_size, max_size)
}
\arguments{
\item{genes_drawn}{Character vector of input genes/proteins of interest}

\item{universe}{Character vector of background genes/proteins}

\item{database}{List of gene sets (e.g. from \code{fgsea::gmtPathways}),
where each element is a character vector of genes}

\item{min_size}{Minimum size of gene sets to include}

\item{max_size}{Maximum size of gene sets to include}
}
\value{
A data frame of enrichment results, including:
\describe{
\item{pathway}{Gene set name}
\item{pvalue}{Raw p-value from hypergeometric test}
\item{pv_bh}{Adjusted p-value (Benjamini-Hochberg)}
\item{gene_ratio}{Proportion of genes overlapping the gene set}
\item{gene_count}{Number of overlapping genes}
\item{set_size}{Size of the gene set}
\item{set_size_in_universe}{Size of the gene set within the universe}
\item{geneset_drawn}{Overlapping genes}
\item{geneset_drawn_in_universe}{Overlapping genes within the universe}
}
}
\description{
Tests whether input genes are enriched in predefined gene sets using a
hypergeometric test, accounting for a specified background universe.
}
\details{
Gene sets are filtered by size, and p-values are adjusted for multiple
testing using the Benjamini-Hochberg method.
}
